requirements.txt
playwright
pytest
pytest-playwright
pytest-xdist
allure-pytest
requests